BC Break: Stop generating foreign keys in models
