+++ example/dataloader/types.json
@@ -1,0 +1,4 @@
+{
+    "Order": "github.com/vektah/gqlgen/example/dataloader.Order",
+    "Customer": "github.com/vektah/gqlgen/example/dataloader.Customer"
+}
